+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.22.0]
 ### Add
 - Integer axes only show integer ticks
+- Handle date as continuous value on axes
 
 ### Fix
 - Add reference_path to all Primitive / Elementary drawing Objects